refactor(goal-detail): split the maturity helpers out of goalDetailShared (#467)
Second slice of the goalDetailShared breakup. Move the bank-deposit maturity
helpers — fmtMaturity (+ the Maturity type), needsMaturityAction, and
needsBookMaturityAction — into a new `goalDetailMaturity.ts`. It imports InvRow as a
type only, so there's no runtime dependency back on goalDetailShared (whose
BankInfoStrip still calls fmtMaturity — now via the new module). Consumers updated:
GoalDetailSheet / DesktopGoalDetail (needs* predicates), MaturityResolveSheet + the
shared test (fmtMaturity). The now-unused @/lib/maturity imports drop out of
goalDetailShared. No behavior change.

Verified: full unit suite 1261 passed (the shared test exercises fmtMaturity from
the new module); typecheck 0 errors; lint 0 errors; git diff --check clean.
goalDetailShared 647 → 592 lines.
