+++ app/assets/components/__tests__/goalDetailShared.test.tsx
@@ -1,5 +1,6 @@
 import { describe, it, expect } from 'vitest'
-import { buildInvRows, buildRenewalSummary, fmtMaturity, type GoalDetailTx } from '../goalDetailShared'
+import { buildInvRows, buildRenewalSummary, type GoalDetailTx } from '../goalDetailShared'
+import { fmtMaturity } from '../goalDetailMaturity'
 import { calcDeadlineMonths, computeGoalCalculator, describeHistoryRow } from '../goalDetailModel'
 import type { FundBreakdownItem } from '../../DashboardClient'
 import { ArrowUpRight, ArrowDownRight, PiggyBank, GitMerge, RefreshCw } from 'lucide-react'
