+++ app/assets/components/goalDetailShared.tsx
@@ -10,7 +10,7 @@ import { formatIntVN, parseIntVN, formatDecimalVN, parseDecimalVN } from '@/lib/
 import { calcProjectedInterest } from '@/lib/finance'
 import { blendedRate } from '@/lib/accumulating'
 import { fmtTxDate } from './transactionUtils'
-import { isTermDeposit, depositMaturityState, isMaturityActionable, isActionableAccumulatingBook } from '@/lib/maturity'
+import { fmtMaturity, type Maturity } from './goalDetailMaturity'
 import type { FundBreakdownItem } from '../DashboardClient'
 
 export const GD_COLORS: Record<string, string> = {
@@ -446,61 +446,6 @@ export function buildInvRows(
   return [...singleRows, ...bookRows]
 }
 
-// A bank-deposit maturity date, formatted for display plus a relative
-// "time left" summary. Returns null when there's no date. `tone` drives the
-// colour: 'neg' once matured (needs action — consistent with the maturity card
-// and resolve pill), 'warn' when due within 30 days (or today), 'neutral'
-// otherwise (issue #263).
-export interface Maturity {
-  formatted: string
-  diffDays: number
-  relative: string
-  tone: 'neutral' | 'warn' | 'pos' | 'neg'
-}
-
-// Whether a holding is a term deposit at (or within the reminder window of) its
-// maturity date — i.e. it needs a renew/withdraw decision. Shared so the mobile
-// sheet and desktop panel surface the "Handle maturity" action identically.
-export function needsMaturityAction(inv: InvRow, isVi: boolean): boolean {
-  if (!isTermDeposit({ type: inv.type, interestRate: inv.interestRate, expiryDate: inv.expiryDate, depositGroupId: inv.depositGroupId })) return false
-  const m = fmtMaturity(inv.expiryDate, isVi)
-  if (!m) return false
-  return isMaturityActionable(depositMaturityState(m.diffDays))
-}
-
-// The book counterpart of needsMaturityAction: a matured/maturing accumulating
-// book needs a book-level collapse decision (the single-row path above excludes
-// grouped rows, so this is the only entry point that surfaces it). Drives the
-// same "Handle maturity" action — MaturityResolveBody branches to collapse for a book.
-export function needsBookMaturityAction(inv: InvRow): boolean {
-  return isActionableAccumulatingBook({ type: inv.type, expiryDate: inv.expiryDate, depositGroupId: inv.depositGroupId })
-}
-
-export function fmtMaturity(dateStr: string | null | undefined, isVi: boolean): Maturity | null {
-  if (!dateStr) return null
-  const d = new Date(dateStr + 'T00:00:00')
-  if (isNaN(d.getTime())) return null
-  const today = new Date()
-  today.setHours(0, 0, 0, 0)
-  const diffDays = Math.round((d.getTime() - today.getTime()) / 86_400_000)
-  const formatted = fmtTxDate(dateStr, isVi ? 'vi' : 'en')
-
-  let relative: string
-  let tone: Maturity['tone'] = 'neutral'
-  if (diffDays < 0) {
-    relative = isVi ? 'Đã đáo hạn' : 'Matured'
-    tone = 'neg'
-  } else if (diffDays === 0) {
-    relative = isVi ? 'Đáo hạn hôm nay' : 'Matures today'
-    tone = 'warn'
-  } else if (diffDays <= 30) {
-    relative = isVi ? `Còn ${diffDays} ngày` : `${diffDays} day${diffDays === 1 ? '' : 's'} left`
-    tone = 'warn'
-  } else {
-    relative = isVi ? `Còn ${diffDays} ngày` : `${diffDays} days left`
-  }
-  return { formatted, diffDays, relative, tone }
-}
 
 // Bank-deposit info strip shown in the investment Options modal: interest rate,
 // maturity date and time-left. Renders nothing for non-bank holdings or when
